requirements.txt
pytest-cov>=2.2.0
boto3
apache-libcloud
openstacksdk>=0.28.0
